{{cookiecutter.project_slug}}/requirements-dev.txt
-r requirements.txt

# Code formatting
black

# Linting
flake8

# Testing with Pytest and pytest-django
pytest
pytest-django
